Update what_we_know_en.md
Various changes, based on what I have read.  I am NOT an immunologist or epidemiologist, and don't always remember where I saw stuff, so DO NOT treat me as more of an expert than you.  I'm just trying to help.

The "80% mild" particularly worries me, as it seems pretty clear that that 80% includes an awful lot of stuff that is way way past what I as a layperson would consider "mild".  I think "mild" for most people is "Not bad enough to take off of work", not "hospitalized but not intubated", and there are a lot of people (Republicans in particular) who are saying, "enh, it's the flu, no big deal".  YES IT IS A BIG DEAL!

It would be nice if we had a number for the % who eventually need hospitalization.

Thank you very much for doing this!